requirements.txt
Django==4.2
djangorestframework
django-cors-headers
python-dotenv
openai
rapidfuzz
redis
django-redis
